==> Scripts_bulkData/Step6_TCGA_P16Signature_Evaluation.Rmd <==
---
title: "Step6_TCGA_P16Signature_Evaluation"
author: "Melissa"
date: "05/07/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
mainDir <- here::here() # the MainDir should be the root of Tumor_BRCA_P53:
source(knitr::purl(file.path(mainDir,"Scripts","scRNAseq","mouse","TCGA","Global_Variables_TCGA.Rmd"), quiet=TRUE))

evalDir= "C:/Users/Melissa S/Desktop/BRCA1_Tumorigenesis/Tumor_BRCA_p53/output/scRNAseq/TCGA/Signature_Evaluation"
```

```{r}
tnscores=readxl::read_xlsx(paste0(RDatadir, "/Metadata_forSurvivalanalysis.xlsx"))
tn = readxl::read_xlsx( paste0(RDatadir, "/Metadata_forSurvivalanalysis_OnlyBRCA_Basal.xlsx"))
scores=readxl::read_xlsx( paste0(SIGdir, "/Signature_Scores_All_PANCANCER_p16_EMT_Apop.xlsx")) %>% as.data.frame()
rownames(scores)=scores$sample
```

```{r}
outputDir2= "C:/Users/Melissa S/Desktop/BRCA1_Tumorigenesis/Tumor_BRCA_p53/output/scRNAseq/TCGA/RData/"

tnscores=readxl::read_xlsx("C:/Users/Melissa S/Desktop/BRCA1_Tumorigenesis/Tumor_BRCA_p53/output/scRNAseq/TCGA/RData/Metadata_forSurvivalanalysis.xlsx")

tn = readxl::read_xlsx("C:/Users/Melissa S/Desktop/BRCA1_Tumorigenesis/Tumor_BRCA_p53/output/scRNAseq/TCGA/RData/Metadata_forSurvivalanalysis_OnlyBRCA_Basal.xlsx")

scores=readxl::read_xlsx( "C:/Users/Melissa S/Desktop/BRCA1_Tumorigenesis/Tumor_BRCA_p53/output/scRNAseq/TCGA/Signatures/Signature_Scores_All_PANCANCER_p16_EMT_Apop.xlsx")%>% as.data.frame()
rownames(scores)=scores$sample

genexp= qs::qread("C:/Users/Melissa S/Desktop/BRCA1_Tumorigenesis/Tumor_BRCA_p53/output/scRNAseq/TCGA/RData/Normalized_Counts_Matreix_BASALvsOthers.qs")
```

```{r}
scores=readxl::read_excel(paste0(figDir, "/Signature_Scores_All_PANCANCER_p16_EMT_Apop.xlsx"))
```

```{r}
tnscores= scores %>% dplyr::filter(scores$sample %in% tn$Sample.ID)
rownames(tnscores)=tnscores$sample
stopifnot(identical(tn$Sample.ID, rownames(tnscores)))

clin= tn
tn$p16Up=tnscores$p16UP_signature_UCell
clin$p16Up=tnscores$p16UP_signature_UCell
writexl::write_xlsx(clin, file.path(figDir, "ClinicalData_PanCancer_Basal.xlsx"))
```



#-------------- WORK ON THE SIGNATURE---------------------------###
```{r}
sigP=p16d
genexp=normalized_counts

gene=intersect(rownames(genexp), toupper(sigP$gene))
mat.exp= genexp[,clin$Sample.ID]
mat.exp=as.data.frame(mat.exp)
```

Recursive feature elimination:
```{r}
com_RMgenes=data.frame("gene_RM"=NA, "pvalue"=NA)
```

Test feature selection metrics:
```{r}
mat.expr=genexp[which(rownames(genexp) %in% gene), ] %>% as.data.frame()
mat.expr=mat.expr[,clin$Sample.ID]
mat.expr=as.data.frame(t(mat.expr))
tn$stage= tn$Neoplasm.Disease.Stage.American.Joint.Committee.on.Cancer.Code

clin$simplifiedStage= ifelse(clin$stage== "STAGE I"| clin$stage== "STAGE IA", "early","late")

clin=as.data.frame(clin)
mat.expr$stage=clin$simplifiedStage
```

```{r}
library(caret)
mat.expr$stage = as.factor(mat.expr$stage)
parts = createDataPartition(mat.expr$stage, p = .8, list = F)

train = mat.expr[parts, ]
test = mat.expr[-parts, ]

X_train = train[,-49]
y_train = train[,49]
y_train= as.factor(y_train)
#-----------------
control_rfe = rfeControl(functions = rfFuncs, # random forest
                      method = "repeatedcv", # repeated cv
                      repeats = 5, # number of repeats
                      number = 10) # number of folds
set.seed(50)
# Performing RFE
library(randomForest)
result_rfe = rfe(x = X_train, 
                 y = y_train, 
                 sizes = c(5:15, 20, 25, 30, 35,40,45,48),
                 rfeControl = control_rfe)

# summarising the results
result_rfe #The top 5 variables (out of 7):
   #NME1, EMB, IGFBP4, HSPE1, VIM

gene.predictors= predictors(result_rfe)
```

```{r}
varimp_data <- data.frame(feature = row.names(varImp(result_rfe)),
                          importance = varImp(result_rfe)[, 1])


g <- ggplot(data = result_rfe, metric = "Accuracy") + theme_bw()
h <- ggplot(data = result_rfe, metric = "Kappa") + theme_bw()

p <- ggplot(data = varimp_data, 
       aes(x = reorder(feature, -importance), y = importance, fill = feature)) +
  geom_bar(stat="identity") + labs(x = "Features", y = "Variable Importance") + 
  geom_text(aes(label = round(importance, 2)), vjust=1, color="black", size=4) + 
  theme_bw() + theme(legend.position = "none") + coord_flip()

qs::qsave(result_rfe, file.path(figDir, "result_rfe_NewData_Revision_VariableImportance_Bulk.qs"))

pdf(paste0(figDir, "VariableImportanceEvaluation_TCGA_TrainingSet2.pdf"))
print(p)
print(g)
print(h)
dev.off()

varimp_data$gene=rownames(varimp_data)
writexl::write_xlsx(varimp_data, paste0(figDir, "VarImportance_DF2_NewData_Revision.xlsx"))
qs::qsave(result_rfe, paste0(RdataDir, "result_rfe_VarImportance2.qs"))
```

```{r}
# Post prediction
postResample(predict(result_rfe, test[,-49]), obs= as.factor(test[,49]))
#Accuracy     Kappa 
#0.9090909 0.3694268 
###  Accuracy       Kappa 0.84848485 -0.05095541 
```

Test with the top 5 or top 15 variables:
```{r}
p16list= list(p16s_5g= c(gene.predictors[1:5]), 
              p16s_6g= c(gene.predictors[1:6]), 
              p16s_7g= c(gene.predictors[1:7]), 
              p16s_8g= c(gene.predictors[1:8]), 
              p16s_9g= c(gene.predictors[1:9]), 
              p16s_10g= c(gene.predictors[1:10]), 
              p16s_11g= gene.predictors[1:11], 
              p16s_12g=gene.predictors[1:12],
              p16s_13g=gene.predictors[1:13],
              p16s_14g=gene.predictors[1:14],
            p16s_15g=gene.predictors[1:15],
            p16s_20g=gene.predictors[1:20],
            p16s_25g=gene.predictors[1:25],
              p16s_30g= gene.predictors[1:30])
              #p16s_35g=gene.predictors[1:35],
            # p16s_40g=gene.predictors[1:40],
            # p16s_45g=gene.predictors[1:45],
            # p16s_50g=gene.predictors[1:50])

p16list= list(p16s_5g= c(gene.predictors[1:7]))

library(data.table)
library(UCell)
dg=as.data.frame(mat.exp)
sig.res2= UCell::ScoreSignatures_UCell(matrix = dg, features= p16list, maxRank = 15000)
sig.res2=as.data.frame(sig.res2)
##
identical(rownames(sig.res2), tn$Sample.ID)

sig.res2$stage= mat.expr$stage

qs::qsave(sig.res2,paste0(figDir, "Signature_P16_AfterFeatureSelection_Basallike_TCGA.qs"))

sig.res= qs::qread(paste0(figDir, "Signature_P16_AfterFeatureSelection_Basallike_TCGA.qs"))
```




TEST ON CPTAC:

```{r}
bsl= cptac_clin %>% filter(PAM50== "Basal")

bsl$overall_survival= as.numeric(as.character(bsl$overall_survival))
bsl$status= as.numeric(as.character(bsl$status))
##
cptac.dt= ge[, which(colnames(ge) %in% rownames(bsl))]



p16list= list(p16s_5g= c(varimp_data$feature[1:5]), 
              p16s_6g= c(varimp_data$feature[1:6]), 
              p16s_7g= c(varimp_data$feature[1:7]), 
              p16s_8g= c(varimp_data$feature[1:8]), 
              p16s_9g= c(varimp_data$feature[1:9]), 
              p16s_10g= c(varimp_data$feature[1:10]), 
              p16s_11g= varimp_data$feature[1:11], 
              p16s_12g=varimp_data$feature[1:12],
              p16s_13g=varimp_data$feature[1:13],
              p16s_14g=varimp_data$feature[1:14],
            p16s_15g=varimp_data$feature[1:15],
            p16s_20g=varimp_data$feature[1:20],
            p16s_25g=varimp_data$feature[1:25],
              p16s_30g= varimp_data$feature[1:30],
              p16s_35g=varimp_data$feature[1:35]
            )
library(data.table)
library(UCell)
cptac.dt=as.data.frame(cptac.dt)
sig.resCPTAC= UCell::ScoreSignatures_UCell(matrix = cptac.dt, features= p16list, maxRank = 15000)
sig.resCPTAC=as.data.frame(sig.resCPTAC)
identical(rownames(sig.resCPTAC), rownames(bsl))
sig.resCPTAC=sig.resCPTAC[rownames(bsl),]
sig.resCPTAC$stage= bsl$pathologic_stage
##---------------
sig.resCPTAC= sig.resCPTAC %>% filter(! is.na(sig.resCPTAC$stage))

sig.resCPTAC$stage= ifelse(sig.resCPTAC$stage== "stagei", "early", "late")
qs::qsave(sig.resCPTAC, paste0(outputDir2, "CPTAC_SignaturesAllgenes.qs"))
```

Work on the signatures:
https://stackoverflow.com/questions/27918320/what-does-negative-incmse-in-randomforest-package-mean

```{r}
resTCGA2= data.frame("Ngene"= colnames(sig.res2)[-ncol(sig.res2)], "pvalue"=NA)
sig.res2$stage= factor(sig.res2$stage, levels = c("early", "late"))

for(i in seq_len(ncol(sig.res2)-1)){
  resTCGA2[i,2]= wilcox.test(sig.res2[,i] ~ sig.res2$stage, alternative= "greater")$p.value
}


#####
resTCGA2$nbr= c(seq(5,15), 20, 25, 30)

p <- resCPTAC  %>% ggplot(aes(x= nbr, y= pvalue)) + geom_line() + themplot + geom_hline(yintercept = 0.05, color= "red") + geom_vline(xintercept = 13, color="blue") + ggtitle("CPTAC_Pvalue_WilcoxStage")

g <- resTCGA2  %>% ggplot(aes(x= nbr, y= pvalue)) + geom_line() + themplot + geom_hline(yintercept = 0.05, color= "red") + geom_vline(xintercept = 25, color="blue") + ggtitle("TCGA_Pvalue_WilcoxStage")

pdf(paste0(RdataDir, "Pvalue_Wilcox_NbrGenes_P16Signature.pdf"))
#print(p)
print(g)
dev.off()
```

```{r}
#---------------------------##
resCPTAC= data.frame("Ngene"= colnames(sig.resCPTAC)[-ncol(sig.resCPTAC)], "pvalue"=NA)

sig.resCPTAC$stage= factor(sig.resCPTAC$stage, levels = c("early", "late"))

for(i in seq_len(ncol(sig.resCPTAC)-1)){
  resCPTAC[i,2]= wilcox.test(sig.resCPTAC[,i] ~ sig.resCPTAC$stage, alternative= "greater")$p.value
}
resCPTAC$nbr= c(seq(5,15), 20, 25, 30, 35)
```

Test on the CPTACT dataset: DONE ON THE CPTAC WORKBOOK

```{r}
l= list("varimp"= varimp_data, "Pval_TCGA"= resTCGA2) # "Pval_CPTAC"= resCPTAC,
writexl::write_xlsx(l, paste0(RdataDir, "List_VarianceImportance_TCGA.xlsx"))
```

##--------------------- SURVIVAL TESTS -------------------###
```{r}
#Try with 5 genes only: "EMB"    "IGFBP4" "EMP3"   "VIM"    "UBE2C" 

```

```{r}
sig.res=sig.res2
identical(tn$Sample.ID, rownames(sig.res))
sig.res$Progress.Free.Survival..Months.=tn$Progress.Free.Survival..Months.

sig.res$Progression.Free.Status_binary=tn$Progression.Free.Status
sig.res$Progression.Free.Status_binary <- ifelse(sig.res$Progression.Free.Status=="0:CENSORED",0,1)

#P16Low should be tn$p16Up < quantile(tn$p16Up, 0.25)
tmpP= data.frame("LengthGenes"= colnames(sig.res)[1:14], "Best_Qtl"= NA, "Pval_Min"=NA) #,1,seq(5, nrow(sigP), 5), nrow(sigP)
rownames(tmpP)= colnames(sig.res)[1:14]

i=0
for(j in seq(1,14)){
  
  qtl=list("quantile"=NA, "hazard"=NA, "waldtest"=NA,"logTestPval"=NA,"waldtestPval"=NA)

for(i in seq(0.1, 0.9, 0.01)){
  sig.res$p16discrete= ifelse(sig.res[,j] > quantile(sig.res[,j], i), 1, 0)
  
   res.cox <- summary(coxph(Surv(Progress.Free.Survival..Months.,Progression.Free.Status_binary ) ~ p16discrete, data = sig.res))
   
  qtl$quantile= append(qtl$quantile, i)
  qtl$hazard=append(qtl$hazard,res.cox$coefficients[,2])
  ##
 qtl$logTestPval= append(qtl$logTestPval, res.cox$logtest[3])
  qtl$waldtest= append(qtl$waldtest,res.cox$waldtest[1] )
   qtl$waldtestPval= append(qtl$waldtestPval,res.cox$waldtest[3] )
} 
  
  
 dt.qtl_binary= do.call(rbind, qtl) %>% t() %>%  as.data.frame()
dt.qtl_binary=dt.qtl_binary[-1,]


bestqtl=  dt.qtl_binary[order( dt.qtl_binary$waldtestPval, decreasing = FALSE),1][1]
bestpval=  dt.qtl_binary[order( dt.qtl_binary$waldtestPval, decreasing = FALSE),5][1]
####
tmpP[j,2]=bestqtl

tmpP[j,3]=bestpval

sig.res$p16discrete= ifelse(sig.res[,j] > quantile(sig.res[,j], bestqtl), "high", "low" )

fit <- survfit(Surv(Progress.Free.Survival..Months.,Progression.Free.Status_binary) ~ p16discrete, data = sig.res)
library(survminer)
pdf(paste0(RdataDir,"SurvivalPlot_TCGA_Signature_Top",tmpP[j,1],"_BestQtl_", bestqtl,"_WaldTest.pdf"), height = 5, width = 5)
 p=ggsurvplot(fit,
          pval = TRUE, conf.int = FALSE,
          risk.table = FALSE, 
          risk.table.col = "strata",
          linetype = "strata", 
          surv.median.line = "hv", 
          ylab="Overall surrival",
          ggtheme = theme_bw(), 
          palette = c("#DED5D5", "#262424", "#756F6F")) 
 print(p)
dev.off()
}

writexl::write_xlsx(tmpP, paste0(RdataDir, "/TCGA_RFE_Survival_WaldTest.xlsx"))

#####################
tmpP$LengthGenes= gsub("p16s_", "", tmpP$LengthGenes)
tmpP$LengthGenes=gsub("g_UCell","", tmpP$LengthGenes)

tmpP$LengthGenes=as.numeric(as.character(tmpP$LengthGenes))

pdf(paste0(RdataDir, "/Lines_TCGA_BestParams_VarImp_WaldTest.pdf"))
m <- tmpP %>% ggplot(aes(x= LengthGenes, y= Pval_Min,group = 1))  + geom_line() + geom_point()+ themplot + xlab("Top_VarImp_RFE")

n <- tmpP %>% ggplot(aes(x= LengthGenes, y= Best_Qtl,group = 1)) + geom_point() + geom_line() + themplot + xlab("Top_VarImp_RFE")

print(m)
print(n)
dev.off()
```
```{r}
tmpP= data.frame("LengthGenes"= colnames(sig.res)[1], "Best_Qtl"= NA, "Pval_Min"=NA) #,1,seq(5, nrow(sigP), 5), nrow(sigP)
rownames(tmpP)= colnames(sig.res)[1]

i=0
for(j in seq(1)){
  
  qtl=list("quantile"=NA, "hazard"=NA, "waldtest"=NA,"logTestPval"=NA,"waldtestPval"=NA)

for(i in seq(0.1, 0.9, 0.01)){
  sig.res$p16discrete= ifelse(sig.res[,j] > quantile(sig.res[,j], i), 1, 0)
  
   res.cox <- summary(coxph(Surv(Progress.Free.Survival..Months.,Progression.Free.Status_binary ) ~ p16discrete, data = sig.res))
   
  qtl$quantile= append(qtl$quantile, i)
  qtl$hazard=append(qtl$hazard,res.cox$coefficients[,2])
  ##
 qtl$logTestPval= append(qtl$logTestPval, res.cox$logtest[3])
  qtl$waldtest= append(qtl$waldtest,res.cox$waldtest[1] )
   qtl$waldtestPval= append(qtl$waldtestPval,res.cox$waldtest[3] )
} 
  
  
 dt.qtl_binary= do.call(rbind, qtl) %>% t() %>%  as.data.frame()
dt.qtl_binary=dt.qtl_binary[-1,]


bestqtl=  dt.qtl_binary[order( dt.qtl_binary$waldtestPval, decreasing = FALSE),1][1]
bestpval=  dt.qtl_binary[order( dt.qtl_binary$waldtestPval, decreasing = FALSE),5][1]
####
tmpP[j,2]=bestqtl

tmpP[j,3]=bestpval

sig.res$p16discrete= ifelse(sig.res[,j] > quantile(sig.res[,j], bestqtl), "high", "low" )

fit <- survfit(Surv(Progress.Free.Survival..Months.,Progression.Free.Status_binary) ~ p16discrete, data = sig.res)
library(survminer)
pdf(paste0(figDir,"SurvivalPlot_TCGA_Signature_Top7genes_",tmpP[j,1],"_BestQtl_", bestqtl,"_WaldTest.pdf"), height = 5, width = 5)
 p=ggsurvplot(fit,
          pval = TRUE, conf.int = FALSE,
          risk.table = FALSE, 
          risk.table.col = "strata",
          linetype = "strata", 
          surv.median.line = "hv", 
          ylab="Overall surrival",
          ggtheme = theme_bw(), 
          palette = c("#DED5D5", "#262424", "#756F6F")) 
 print(p)
dev.off()
}

writexl::write_xlsx(tmpP, paste0(RdataDir, "/TCGA_RFE_Survival_WaldTest.xlsx"))

#####################
tmpP$LengthGenes= gsub("p16s_", "", tmpP$LengthGenes)
tmpP$LengthGenes=gsub("g_UCell","", tmpP$LengthGenes)

tmpP$LengthGenes=as.numeric(as.character(tmpP$LengthGenes))

pdf(paste0(RdataDir, "/Lines_TCGA_BestParams_VarImp_WaldTest.pdf"))
m <- tmpP %>% ggplot(aes(x= LengthGenes, y= Pval_Min,group = 1))  + geom_line() + geom_point()+ themplot + xlab("Top_VarImp_RFE")

n <- tmpP %>% ggplot(aes(x= LengthGenes, y= Best_Qtl,group = 1)) + geom_point() + geom_line() + themplot + xlab("Top_VarImp_RFE")

print(m)
print(n)
dev.off()
```




for the paper:
```{r}
identical(tn$Sample.ID, rownames(sig.res))
sig.res$Progress.Free.Survival..Months.=tn$Progress.Free.Survival..Months.

sig.res$Progression.Free.Status_binary=tn$Progression.Free.Status
sig.res$Progression.Free.Status_binary <- ifelse(sig.res$Progression.Free.Status=="0:CENSORED",0,1)

#P16Low should be tn$p16Up < quantile(tn$p16Up, 0.25)
#tmpP= data.frame("LengthGenes"= colnames(sig.res)[1:16], "Best_Qtl"= NA, "Pval_Min"=NA) #,1,seq(5, nrow(sigP), 5), nrow(sigP)
#rownames(tmpP)= colnames(sig.res)[1:16]
#i=0
#for(j in seq(1,16)){
  
  qtl=list("quantile"=NA, "hazard"=NA, "waldtest"=NA,"logTestPval"=NA,"waldtestPval"=NA)

for(i in seq(0.1, 0.9, 0.01)){
  sig.res$p16discrete= ifelse(sig.res[,9] > quantile(sig.res[,9], i), 1, 0)
  
   res.cox <- summary(coxph(Surv(Progress.Free.Survival..Months.,Progression.Free.Status_binary ) ~ p16discrete, data = sig.res))
   
  qtl$quantile= append(qtl$quantile, i)
  qtl$hazard=append(qtl$hazard,res.cox$coefficients[,2])
  ##
 qtl$logTestPval= append(qtl$logTestPval, res.cox$logtest[3])
  qtl$waldtest= append(qtl$waldtest,res.cox$waldtest[1] )
   qtl$waldtestPval= append(qtl$waldtestPval,res.cox$waldtest[3] )
} 
  
  
 dt.qtl_binary= do.call(rbind, qtl) %>% t() %>%  as.data.frame()
dt.qtl_binary=dt.qtl_binary[-1,]

qs::qsave(dt.qtl_binary, paste0(outputDir2, "/TCGA_Best_Qtl_preHumanSignature_13genes.qs"))



pdf(paste0(outputDir2, "BestQtl_Pvalue_preHuman13gene_WaldTest.pdf"))
m <- dt.qtl_binary %>% ggplot(aes(x= quantile, y= waldtestPval)) + geom_bar(stat="identity") + themplot
print(m)
dev.off()

```


BoxPlot :

```{r}
pdf(paste0(RdataDir, "/Boxplot_Top13g_38g_30g_TCGA.pdf"))
p <- ggplot(data= sig.res2, aes(x=stage, y= p16s_13g_UCell)) + geom_boxplot()  + themplot + geom_jitter(shape=16, position=position_jitter(0.2), alpha=0.5) + ggtitle("Important_13genes") 

g <- ggplot(data= sig.res2, aes(x=stage, y= p16s_25g_UCell)) + geom_boxplot()  + themplot + geom_jitter(shape=16, position=position_jitter(0.2), alpha=0.5) + ggtitle("Important_38genes")

h <- ggplot(data= sig.res2, aes(x=stage, y= p16s_30g_UCell)) + geom_boxplot()  + themplot + geom_jitter(shape=16, position=position_jitter(0.2), alpha=0.5) + ggtitle("Important_30genes")
print(p)
print(h)
print(g)
dev.off()
```

For the paper: 25 as it showed the lowest p value for predicting early stage
```{r}
varimp_data$i= ifelse(as.numeric(varimp_data$gene) < 26, "Important", "NotImportant")
pdf(paste0(RdataDir, "Barplot_VarImportance_ForPaper.pdf"))
p <- ggplot(data = varimp_data, 
       aes(x = reorder(feature, -importance), y = importance, fill = i)) +
  geom_bar(stat="identity") + labs(x = "Features", y = "Variable Importance")+ 
  themplot + theme(legend.position = "none", axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) + scale_fill_manual(values=c("Important"= "black", "NotImportant"="grey"))
print(p)
dev.off()
```
