==> Scripts_bulkData/Step7_TCGA_Signature_Association_Survival.Rmd <==
---
title: "Step7_TCGA_Signature_Association_Survival"
author: "Melissa"
date: "13/07/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
mainDir <- here::here() # the MainDir should be the root of Tumor_BRCA_P53:
source(knitr::purl(file.path(mainDir,"Scripts","scRNAseq","mouse","TCGA","Global_Variables_TCGA.Rmd"), quiet=TRUE))

evalDir= "C:/Users/Melissa S/Desktop/BRCA1_Tumorigenesis/Tumor_BRCA_p53/output/scRNAseq/TCGA/Signature_Evaluation"
```

Check the variables colinearity:
```{r}
#Keep only positive:
basal <- log_norm.matB[which(rownames(log_norm.matB) %in% p16d$gene),] %>% t()

library(corrplot)
corrplot(cor(basal))
library(caret)
indexesToDrop <- findCorrelation(cor(basal), cutoff = 0.7) #"EMP3"  "SPARC"
basal=basal[,-indexesToDrop]
basal=as.data.frame(basal)
```

## Predict the stage of the disease:
```{r}
stopifnot(identical(tn$Sample.ID, rownames(basal)))
basal$stage= tn$simplifiedStage
basal$time= tn$Disease.Free..Months.
basal$status= tn$Disease.Free.Status
basal$binary_stage= ifelse(tn$stage=="1", "early", "late")
```

Prepare the data:
```{r}
train_basal= basal[sample(rownames(basal), size= nrow(basal) * 0.7, replace = FALSE),]
test_basal= basal[-which(rownames(basal) %in% rownames(train_basal)),]
```

##--------------------- GLMNET------------------------####
Train and test the model :
```{r}
lambda_seq <- 10^seq(2, -2, by = -.1)
set.seed(86)
cv_output <- cv.glmnet(x= as.matrix(train_basal[,- seq(ncol(train_basal)-3, ncol(train_basal))]), y= as.factor(train_basal$binary_stage),
                       alpha = 1, lambda = lambda_seq, 
                       nfolds = 5, family=binomial(link = "logit"))
best_lam <- cv_output$lambda.min

# Rebuilding the model with best lamda value identified
lasso_best <- glmnet(x= as.matrix(train_basal[,- seq(ncol(train_basal)-3, ncol(train_basal))]), y= as.factor(train_basal$binary_stage), alpha = 1, lambda = best_lam, family=binomial(link = "logit"))
#get the associated parameters:
weights <- coef(lasso_best) %>% as.matrix() %>% as.data.frame()




pred <- predict(lasso_best, s = best_lam, newx = as.matrix(test_basal[,- seq(ncol(train_basal)-3, ncol(train_basal))]))
```

```{r}
sds <- apply(train_basal[,- seq(ncol(train_basal)-3, ncol(train_basal))], 2, sd)
cs <- as.matrix(coef(lasso_best, s = "lambda.min"))
std_coefs <- cs[-1, 1] * sds
```

##--------- TEST LOGISTIC REGRESSION USING GLM---------------- ##
```{r}
model <- glm( as.factor(binary_stage) ~., data = train_basal, family = "binomial")

summary(model)$coef

model2 <-  glmnet(x= as.matrix(train_basal[,- ncol(train_basal)]), y= as.factor(train_basal$binary_stage), alpha = 1,  family=binomial(link = "logit"))
```

















































#-------------- STEP2 ---------------------------------------------####
Try to filter out genes associated with prognosis:

Cox model on eacg gene:
```{r}
tmp1= genexp[,md$Sample.ID]
tmp1= tmp1[which(rownames(tmp1) %in% c(toupper(sigP$gene), toupper(sigN$gene))),]
stopifnot(identical(colnames(tmp1), md$Sample.ID))
md= cbind(md, as.data.frame(t(tmp1)))
###
dt.= data.frame("gene"= rownames(tmp1), "Pvalue_WaldTest"= NA, "Pvalue_LogRankTest"=NA)
for(i in seq_len(nrow(tmp1))){
dt.[i,2] <- summary(coxph(Surv(Disease.Free..Months.,Progression.Free.Status_binary ) ~ tmp1[i,], data = md))$waldtest[3]
dt.[i,3] <- summary(coxph(Surv(Disease.Free..Months.,Progression.Free.Status_binary ) ~ tmp1[i,], data = md))$logtest[3]

}
#only for pos:
stopifnot(identical(dt.$gene, toupper(sigP$gene)))
sigP=sigP[match(Hmisc::capitalize(tolower(dt.$gene)), sigP$gene),]
dt.$logFC=sigP$avg_log2FC


writexl::write_xlsx(dt., paste0(evalDir, "/CoxRegression_TopPositive_NegativeGenes.xlsx"))
dt. <- readxl::read_excel(paste0(evalDir, "/CoxRegression_TopPositive_NegativeGenes.xlsx"))

dt.$keep= ifelse(dt.$Pvalue_LogRankTest < 0.05, "Yes", "No")

```



```{r}
clin=tn[,-c(100:103)]

tmp1= genexp[,clin$Sample.ID]
tmp1= tmp1[which(rownames(tmp1) %in% gn$gene),]

stopifnot(identical(colnames(tmp1), clin$Sample.ID))
clin= cbind(clin, as.data.frame(t(tmp1)))

clin$Progression.Free.Status_binary <- ifelse(clin$Progression.Free.Status=="0:CENSORED",0,1)
 
md <- clin %>% select(c("Disease.Free..Months.","Progression.Free.Status_binary", rownames(tmp1)))
```



Test the lasso regression with COX model: DIDNT WORK DUE TO COLINEARITY
```{r}
library(glmnet)
library(survival)
md=tn

md = md %>% filter(!is.na(Disease.Free..Months.)==TRUE)
md$time= md$Disease.Free..Months.
md$status=md$Disease.Free.Status
md$status= ifelse(md$status=="0:DiseaseFree", 0, 1)
#colnames(md)= c("time", "status", colnames(md)[3:ncol(md)])
#md= md %>% filter(time > 0)
identical(rownames(basal2), md$Sample.ID)
basal2= basal2[md$Sample.ID,]
basal2= cbind(basal2, md[, c("time", "status")])
basal2= basal2 %>% filter(time > 0)
so <-  Surv(basal2$time,basal2$status )

fit <- glmnet(x= basal2[, -c(76, 75)], y=so, family = "cox")

#bsl2= norm.matB[, rownames(basal2)]
bsl2= norm.matB
bsl2=as.data.frame(t(bsl2))

identical(rownames(bsl2), rownames(basal2))

#bsl2 <- as(as.matrix(bsl2), "sparseMatrix")

fitcv= cv.glmnet(x= as.matrix(bsl2), y=so, family = "cox")
#find optimal lambda value that minimizes test MSE
best_lambda <- fitcv$lambda.min
best_lambda # 0.08455911

#produce plot of test MSE by lambda value
pdf(paste0(evalDir, "/Plot_CV.pdf"))
plot(fitcv) 
dev.off()
####

#find coefficients of best model
best_model <- glmnet(x= as.matrix(bsl2), y=so, lambda = best_lambda , family = "cox")
coef(best_model)
###
mat <- coef(best_model) %>% as.data.frame()
mat$s0= -1 *mat$s0
writexl::write_xlsx(mat, paste0(evalDir, "/Lasso_Model_BetaParam.xlsx"))

mat <- readxl::read_excel(paste0(evalDir, "/Lasso_Model_BetaParam.xlsx"))


stopifnot(identical(rownames(mat), colnames(md)[-c(1,2)]))
for(i in seq_len(nrow(md))){
  
  md$signature[i]= mat$s0[1] * md[i,3] + 
    mat$s0[2] * md[i,4] + 
    mat$s0[3] * md[i,5] +
    mat$s0[4] * md[i,6]+ 
    mat$s0[5] * md[i,7] + 
    mat$s0[6] * md[i,8] + 
    mat$s0[7] * md[i,9] +
    mat$s0[8] * md[i,10] 
}
 
md$bin= ifelse(md$signature < median(md$signature), "Neg", "Pos")

writexl::write_xlsx(md, paste0(evalDir, "/LassoSignature_Binarized.xlsx"))

fit <- survfit(Surv(time,status) ~ bin, data = md)
library(survminer)
pdf(paste0(evalDir,"/SurvivalPlot_SignificantGenes_", gn[i,],".pdf"), height = 5, width = 5)
 p=ggsurvplot(fit,
          pval = TRUE, conf.int = FALSE,
          risk.table = FALSE, 
          risk.table.col = "strata",
          linetype = "strata", 
          surv.median.line = "hv", 
          ylab=paste0("RFS", gn[i,]),
          ggtheme = theme_bw()) 
 print(p)
dev.off()
```
